refactor(ui): divide testes de endereço em arquivos específicos
- Divide testes em arquivos por categoria: endereco.basic.cy.js (básicos), endereco.validation.cy.js (validações), endereco.usability.cy.js (usabilidade), endereco.accessibility.cy.js (acessibilidade), endereco.support.js (constantes)

- Atualiza relatório com resultados dos testes

- Atualiza estrutura do projeto no README
