+++ README.md
@@ -8,8 +8,21 @@ Este projeto contém testes automatizados utilizando Cypress para validar o flux
 .
 ├── cypress/
 │   ├── e2e/              
-│   │   ├── api-carrinho.cy.js  
-│   │   └── fluxo-endereco.cy.js      
+│   │   ├── api/                    # Testes de API
+│   │   │   ├── carrinho.basic.cy.js
+│   │   │   ├── carrinho.validation.cy.js
+│   │   │   ├── carrinho.error.cy.js
+│   │   │   ├── carrinho.performance.cy.js
+│   │   │   ├── carrinho.advanced.cy.js
+│   │   │   ├── carrinho.create.cy.js
+│   │   │   ├── carrinho.delete.cy.js
+│   │   │   └── carrinho.support.js
+│   │   └── ui/                     # Testes de Interface
+│   │       ├── endereco.basic.cy.js
+│   │       ├── endereco.validation.cy.js
+│   │       ├── endereco.usability.cy.js
+│   │       ├── endereco.accessibility.cy.js
+│   │       └── endereco.support.js
 │   ├── support/        
 │   └── results/        
 ├── docs/
@@ -44,8 +57,8 @@ npx cypress run
 npx cypress open
 
 # Executar testes específicos
-npx cypress run --spec "cypress/e2e/api-carrinho.cy.js"  # Apenas testes de API
-npx cypress run --spec "cypress/e2e/fluxo-endereco.cy.js"  # Apenas testes de UI
+npx cypress run --spec "cypress/e2e/api*.cy.js"  # Apenas testes de API
+npx cypress run --spec "cypress/e2e/ui*.cy.js"  # Apenas testes de UI
 
 # Abrir Cypress Test Runner (interface gráfica)
 npx cypress open
@@ -72,7 +85,9 @@ Os relatórios de execução dos testes estão disponíveis em:
 - Testes de API: 
    - [`cypress/e2e/api/*.cy.js`](cypress/e2e/api/) 
    - [`docs/RelatorioAPI.md`](docs/RelatorioAPI.md)
-- Testes de UI: [`docs/RelatorioUI.md`](docs/RelatorioUI.md)
+- Testes de UI: 
+   - [`cypress/e2e/ui/*.cy.js`](cypress/e2e/ui/) 
+   - [`docs/RelatorioUI.md`](docs/RelatorioUI.md)
 
 ## CI/CD
